refactor: Updated folder structure
